// src/modules/wallet/domain/entities/external-validation.types.ts
export enum ExternalValidationResponseCode {
  AUTHORIZED = 'AUTHORIZED',
  DECLINED = 'DECLINED',
  DECLINED_INSUFFICIENT_FUNDS = 'DECLINED_INSUFFICIENT_FUNDS',
  DECLINED_LOCAL_CURRENCY_INVALID = 'DECLINED_LOCAL_CURRENCY_INVALID',
  DECLINED_DATETIME_INVALID = 'DECLINED_DATETIME_INVALID',
  DECLINED_CARD_UNKNOW = 'DECLINED_CARD_UNKNOW',
  DECLINED_MCC_INVALID = 'DECLINED_MCC_INVALID',
  DECLINED_MERCHANTID_INVALID = 'DECLINED_MERCHANTID_INVALID',
  DECLINED_MERCHANT_CITY_INVALID = 'DECLINED_MERCHANT_CITY_INVALID',
  DECLINED_MERCHANT_COUNTRY_INVALID = 'DECLINED_MERCHANT_COUNTRY_INVALID',
}
export const externalValidationResponseEnumName =
  'external_validation_response_enum'

export const declinedExternalValidationCode = [
  ExternalValidationResponseCode.DECLINED,
  ExternalValidationResponseCode.DECLINED_INSUFFICIENT_FUNDS,
  ExternalValidationResponseCode.DECLINED_LOCAL_CURRENCY_INVALID,
  ExternalValidationResponseCode.DECLINED_DATETIME_INVALID,
  ExternalValidationResponseCode.DECLINED_CARD_UNKNOW,
  ExternalValidationResponseCode.DECLINED_MCC_INVALID,
  ExternalValidationResponseCode.DECLINED_MERCHANTID_INVALID,
  ExternalValidationResponseCode.DECLINED_MERCHANT_CITY_INVALID,
  ExternalValidationResponseCode.DECLINED_MERCHANT_COUNTRY_INVALID,
]